Looks a lot cleaner. Refactored some code to make it more compartmentalized
